# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Tether-ERP Procurement API
  description: |
    Comprehensive REST API for procurement management system with approval workflows,
    budget constraints, document linking, and multi-stage approval chains.
  version: 1.0.0
  contact:
    name: Tether-ERP Support
    email: support@tether-erp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development Server
  - url: https://api.tether-erp.com/api/v1
    description: Production Server
...
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: List notifications
      operationId: getNotifications
      security:
        - BearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
...
  /notifications/stats:
    get:
      tags:
        - Notifications
      summary: Get notification statistics
      operationId: getNotificationStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Notification stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      unread:
                        type: integer
                      total:
                        type: integer
...
  /notifications/mark-all-as-read:
    post:
      tags:
        - Notifications
      summary: Mark all notifications as read
      operationId: markAllNotificationsAsRead
      security:
        - BearerAuth: []
      responses:
        "200":
          description: All notifications marked as read
  /notifications/{id}:
    delete:
      tags:
        - Notifications
      summary: Delete notification
      operationId: deleteNotification
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        data:
          type: object
        read:
          type: boolean
        readAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Resource not found